Split `helpers.ts` into simpler component oriented modules. Split the helper tests into component oriented test modules.
